% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.odds.ratio.R
\name{common.odds.ratio}
\alias{common.odds.ratio}
\title{Common Odds Ratio of a 2x2xk Table}
\usage{
common.odds.ratio(x)
}
\arguments{
\item{x}{A 3 dimensional 2x2xk table/array of 2x2 tables.}
}
\value{
The common odds ratio of the 2x2xk table.
}
\description{
Calculates the common odds ratio of a 2x2xk table.
}
\examples{
a <- array(c(1,3,5,6,
             2,7,3,1,
             2,4,9,2),
           dim = c(2,2,3))
common.odds.ratio(a)
}
