% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantelhaenzel.test.R
\name{mantelhaenzel.test}
\alias{mantelhaenzel.test}
\title{Cochran–Mantel–Haenszel Chi-Squared Test for a 2x2xk Table.}
\usage{
mantelhaenzel.test(x)
}
\arguments{
\item{x}{A 3 dimensional 2x2xk table/array of 2x2 tables.}
}
\value{
A list including the common odds ratio, chi-squared statistic, and p-value from the test.
}
\description{
Performs the Cochran–Mantel–Haenszel chi-squared test without a continuity correction on a 2x2xk table, testing the null hypothesis that the common odds ratio is 1 and the alternative that it is non-zero.
}
\examples{
a <- array(c(1,3,5,6,
             2,7,3,1,
             2,4,9,2),
           dim = c(2,2,3))
mantelhaenzel.test(a)
}
